pyproject.toml
[tool.poetry]
name = "cas_eresearch_gitlab_app"
version = "0.0.0-dev" # Just a place holder; package version is managed by the dynamic-versioning plugin & git tags
description = "CAS-eRessearch GitLab App"
authors = [ "Poole, Gregory B." ]
license = "MIT-expat"
readme = "README.md"
packages = [
    { include = "cas_eresearch_gitlab_app", from = "python" }
]
homepage = "https://github.com/gbpoole/CASeResearch-gitlab"
documentation = "https://CASeResearch-gitlab.readthedocs.io/en/latest/"

[tool.poetry.dependencies]
python = " >=3.11"
click = "^8.1.3"
Sphinx = { version = "^6.1.3", optional = true }
sphinx-rtd-theme = { version = "1.2.0", optional = true }
myst-parser = { version = "^1.0.0", optional = true }
pre-commit = { version = "^3.0.4", optional = true }
pytest = { version = "^7.0", optional = true }
pytest-cov = { version = "^4.1.0", optional = true }
black = { version = "^22.10.0", optional = true }
ruff = { version = "^0.0.243", optional = true}
sphinx-click = { version = "^4.4.0", optional = true }
mypy = { version = "^1.1.1", optional = true }
sphinx-copybutton = { version = "^0.5.1", optional = true }
pytest-mock = "^3.11.1"
uvicorn = "^0.24.0.post1"
fastapi = "^0.105.0"
httpx = "^0.25.2"
python-decouple = "^3.8"
sqlalchemy = "^2.0.25"
pandas = "^2.2.0"


[tool.poetry.extras]
docs = [
   "Sphinx",
   "sphinx-rtd-theme",
   "myst-parser",
   "sphinx-click",
   "sphinx-copybutton",
]
dev = ["pre-commit",
       "pytest",
       "pytest-cov",
       "mypy",
       "black",
       "ruff",
]

[tool.poetry.scripts]
cas_eresearch_gitlab_app = "cas_eresearch_gitlab_app.cli:cli"

[tool.pytest.ini_options]
addopts = "--cov"
pythonpath = "./python" # relative to root directory
# Uncomment and edit the following line to add custom markers to your tests
# These can be run with the following (for example):
#
#   $ pytest -m marker_name
#
#markers = [ "marker_name: A short description of the marker" ]

# This configures the Poetry plugin which enables dynamic versioning via the repo tags
# Note the changes that are also needed to the [build-system] config below.
[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.mypy]
ignore_missing_imports = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/python"

[tool.coverage.paths]
source = ["python"]

[tool.coverage.run]
branch = true
source = ["python"]
omit = ["python/cas_eresearch_gitlab_app/tests/*"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "if typing.TYPE_CHECKING:"
]

[tool.ruff]
line-length = 88 # to match Black
select = [
    "E",
    "F",
    "W",
]
ignore = [
    "E501", # line length violations (let Black handle this)
]

# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]

[tool.ruff.pydocstyle]
convention = "numpy"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
